+++ Pandora/Paradigm/Basis.hs
@@ -19,7 +19,7 @@ import Pandora.Paradigm.Basis.Constant as Exports
 import Pandora.Paradigm.Basis.Identity as Exports
 
 import Pandora.Core.Functor (type (~>))
-import Pandora.Core.Morphism ((.), (!))
+import Pandora.Core.Morphism ((!))
 
 note :: e -> Maybe ~> Conclusion e
 note x = maybe (Failure x) Success