+++ CHANGELOG.md
@@ -30,3 +30,4 @@
 
 # 0.1.4
 * Define `Jet` datatype in `Basis` module
+* Add `fail` method for `Conclusion` datatype