+++ CHANGELOG.md
@@ -32,3 +32,4 @@
 * Define `Jet` datatype in `Basis` module
 * Add `fail` method for `Conclusion` datatype
 * Define `find` method in terms of stateful traversing
+* Define `filter` method for `Stack` datastructure