+++ CHANGELOG.md
@@ -35,3 +35,5 @@
 * Define `filter` method for `Stack` datastructure
 * Define `loeb` method for `Covariant` type class
 * Define `Variation` datatype in `Basis` module
+* Define infix versions of `comap` with various nesting levels
+* Define infix versions of `contramap` with various nesting levels